src/components/script/dom/bindings/codegen/test/test_InstanceOf.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=748983
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 748983</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=748983">Mozilla Bug 748983</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 748983 **/
ok(document instanceof EventTarget, "document is an event target")
ok(new XMLHttpRequest() instanceof XMLHttpRequest, "instanceof should work on XHR");

</script>
</pre>
</body>
</html>
